requirements.txt
Django==3.0.5
djangorestframework==3.12.1
django-filter==2.4.0